removed unnecessary abstraction and complexity
